+++ src/hkimjp/konpy2/login.clj
@@ -32,10 +32,11 @@
 (defn login!
   [{{:keys [login password]} :params}]
   ;; always login success when (env :auth) is empty
+  (t/log! {:level :debug :id "login!" :msg (str login " " password)})
   (if (empty? (env :auth))
     (do
       (t/log! :info (str "no auth mode: " login))
-      (-> (resp/redirect "/")
+      (-> (resp/redirect "/k/tasks")
           (assoc-in [:session :identity] login)))
     (try
       (let [resp (hc/get (str l22 "/api/user/" login) {:timeout 3000 :as :json})]
@@ -43,20 +44,20 @@
                  (hashers/check password (get-in resp [:body :password])))
           (do
             (t/log! :info (str "login success: " login))
-            (-> (resp/redirect "/")
+            (-> (resp/redirect "/k/tasks")
                 (assoc-in [:session :identity] login)))
           (do
             (t/log! :info (str "login failed: " login))
-            (-> (resp/redirect "/login")
+            (-> (resp/redirect "/")
                 (assoc :session {} :flash "login failed")))))
       ;; happens?
       (catch Exception e
         (t/log! :warn (.getMessage e))
-        (-> (resp/redirect "/login")
+        (-> (resp/redirect "/")
             (assoc :session {} :flash "enter login/password"))))))
 
 (defn logout!
   [request]
   (t/log! :info (str "logout! " (user request)))
-  (-> (resp/redirect "/login")
+  (-> (resp/redirect "/")
       (assoc :session {})))