+++ src/hkimjp/konpy2/middleware.clj
@@ -8,6 +8,7 @@
 (defn wrap-users
   [handler]
   (fn [request]
+    (println (:session request))
     (let [user (user request)]
       (t/log! :debug (str "wrap-users " user))
       (if (some? user)
@@ -16,7 +17,7 @@
           (handler request))
         (do
           (t/log! :debug "not found")
-          (-> (resp/redirect "/login")
+          (-> (resp/redirect "/")
               (assoc :session {} :flash "need login")))))))
 
 (defn wrap-admin [handler]
@@ -25,5 +26,5 @@
       (t/log! :debug (str "wrap-admin " user))
       (if (= (env :admin) user)
         (handler request)
-        (-> (resp/redirect "/login")
+        (-> (resp/redirect "/")
             (assoc :session {} :flash "admin only"))))))