+++ src/hkimjp/konpy2/response.clj
@@ -4,16 +4,16 @@
    [ring.util.response :as resp]
    [taoensso.telemere :as t]))
 
-(def version "0.2.2")
+(def version "0.2.4")
 
 (def ^:private menu "text-xl font-medium text-white px-1 hover:bg-red-500")
 
 (defn navbar []
   [:div.flex.bg-red-600.items-baseline.gap-x-4
    [:div.text-2xl.font-medium.text-white "KONPY2"]
-   [:div {:class menu} [:a {:href "/"} "tasks"]]
-   [:div {:class menu} [:a {:href "/scores"} "scores"]]
-   [:div {:class menu} [:a {:href "/stocks"} "stocks"]]
+   [:div {:class menu} [:a {:href "/k/tasks"} "tasks"]]
+   [:div {:class menu} [:a {:href "/k/scores"} "scores"]]
+   [:div {:class menu} [:a {:href "/k/stocks"} "stocks"]]
    [:div {:class menu} [:a {:href "/logout"} "logout"]]
    [:div {:class menu} [:a {:href "/help"} "HELP"]]
    [:div {:class menu} [:a {:href "/admin"} "(admin)"]]])